pyproject.toml
[project]
name = "sectionprops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.20.0",
    "ruff>=0.11.13",
    "sectionproperties>=3.9.0",
]
